Avoid a signed/unsigned comparison

+++ code/ship/ship.cpp
@@ -14339,7 +14339,7 @@ static void ship_add_ship_type_kill_count( int ship_info_index )
 	}
 
 	//Add it if we are actually in gameplay
-	if (Ship_type_counts.size() > type)
+	if (Ship_type_counts.size() > static_cast<size_t>(type))
 		Ship_type_counts[type].killed++;
 }
